LPS-169164 Avoid synchronization issues
I'm moving all the state to an AtomicReference so that it can never be read and
show an inconsistent content.

I don't want to use synchronized{}, nor ReadWriteLock for performance reasons and
because the possibility of finding state values in an inconsistent state, even if
this synchronization was missing, is almost insignificant.